requirements.txt
fastapi==0.75.2
uvicorn==0.17.6
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
requests==2.27.1
pytest==7.1.2
httpx==0.23.0
python-dotenv==0.20.0